% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{as_data_time}
\alias{as_data_time}
\title{Time format units
This code transforms the time from milliseconds to a calendar date format.}
\usage{
as_data_time(x)
}
\description{
Time format units
This code transforms the time from milliseconds to a calendar date format.
}
\keyword{internal}
