% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/providers.R
\name{get_data_sources}
\alias{get_data_sources}
\title{List providers, description, year and link}
\usage{
get_data_sources(query = NULL)
}
\arguments{
\item{query}{A string. Default is NULL. Filter by provider name(s). For valid values use \code{get_providers()}.}
}
\value{
A tibble object.
}
\description{
List the available providers of the geoidep package.
}
\examples{
\dontrun{
library(geoidep)
get_data_sources()
}
}
