% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_fores_fire_link}
\alias{get_fores_fire_link}
\title{Geobosque API to get deforestation hot-spots for the last week}
\usage{
get_fores_fire_link(type = NULL)
}
\arguments{
\item{type}{A string. Only one layer; `warning_last_week`}
}
\value{
A string containing the URL of the requested file.
}
\description{
Geobosque API to get deforestation hot-spots for the last week
}
\keyword{internal}
