% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_forest_fire_link}
\alias{get_forest_fire_link}
\title{Serfor API to get forest fire}
\usage{
get_forest_fire_link(type = NULL)
}
\arguments{
\item{type}{A string. Only one layer; `forest_fire`}
}
\value{
A string containing the URL of the requested file.
}
\description{
Serfor API to get forest fire
}
\keyword{internal}
