% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_geobosque_link}
\alias{get_geobosque_link}
\title{Geobosque links for obtaining cartographic information about loss and g}
\usage{
get_geobosque_link(type = NULL)
}
\arguments{
\item{type}{A string. Select only one of the following layers; 'bpdist', 'bppro', 'bpdep'. Defaults to NULL.}
}
\value{
A string containing the URL of the requested file.
}
\description{
Geobosque links for obtaining cartographic information about loss and g
}
\keyword{internal}
