% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_hazard_data.R
\name{get_hazard_data}
\alias{get_hazard_data}
\title{Downloading available hazard data from the SIGRID platform}
\usage{
get_hazard_data(layer = NULL, dsn = NULL, show_progress = TRUE, quiet = TRUE)
}
\arguments{
\item{layer}{Select only one from the list of available layers, for more information please use `get_data_sources(provider = "SIGRID")`. Defaults to NULL.}

\item{dsn}{Character. Output filename with the \bold{spatial format}. If missing, a temporary file is created.}

\item{show_progress}{Logical. Suppress bar progress.}

\item{quiet}{Logical. Suppress info message.}
}
\value{
An sf object.
}
\description{
This function allows you to download the latest version of the harzards data available on the SIGRID geoportal.
For more information, you can visit the following web page: \href{https://sigrid.cenepred.gob.pe/sigridv3/}{SIGRID Geoportal}
}
\examples{
\donttest{
library(geoidep)
library(sf)
inundaciones <- get_hazard_data(layer = "inundacion_inventario" , show_progress = FALSE)
head(inundaciones)
plot(st_geometry(inundaciones))
}
}
