% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_heat_spot_link}
\alias{get_heat_spot_link}
\title{Serfor API to get heat spot}
\usage{
get_heat_spot_link(type = NULL)
}
\arguments{
\item{type}{A string. Only one layer; `heat_spot`}
}
\value{
A string containing the URL of the requested file.
}
\description{
Serfor API to get heat spot
}
\keyword{internal}
