% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inaigem.R
\name{get_inaigem_data}
\alias{get_inaigem_data}
\title{Download the available data from INAIGEM}
\usage{
get_inaigem_data(
  layer = NULL,
  dsn = NULL,
  show_progress = TRUE,
  quiet = TRUE,
  timeout = 60
)
}
\arguments{
\item{layer}{Select only one from the list of available layers, for more information please use \code{get_data_sources(provider = "INAIGEM")}. Defaults to NULL.}

\item{dsn}{Character. Output filename. If missing, a temporary file is created.}

\item{show_progress}{Logical. Show a cli progress bar. Default \code{TRUE}.}

\item{quiet}{Logical. Suppress info message. Default \code{TRUE}.}

\item{timeout}{Numeric. Seconds to wait for a response. Default 60.}
}
\value{
An sf object.
}
\description{
Download the latest version of data available on the INAIGEM geoportal.
For more information, visit \href{https://geoportal.inaigem.gob.pe/}{INAIGEM Geoportal}.
}
\examples{
\dontrun{
library(geoidep)
library(sf)
glaciar <- get_inaigem_data(layer = "glaciares_1989", show_progress = FALSE)
head(glaciar)
plot(st_geometry(glaciar))
}
}
