% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_midagri_data.R
\name{get_midagri_data}
\alias{get_midagri_data}
\title{Download the geographic information from MIDAGRI}
\usage{
get_midagri_data(layer = NULL, dsn = NULL, show_progress = TRUE, quiet = TRUE)
}
\arguments{
\item{layer}{Character. Currently, Has three layer available, please review in details sections.}

\item{dsn}{Character. The output filename in \bold{*.shp} format. If not provided, a temporary file will be created.}

\item{show_progress}{Logical. If TRUE, displays a progress bar.}

\item{quiet}{Logical. If TRUE, suppresses informational messages.}
}
\value{
An `sf` object containing the downloaded geographic data.
}
\description{
This function downloads the latest version of the \bold{geographic data} representing the \bold{agricultural area} of Peru.
For more details, please visit the following page: \url{https://siea.midagri.gob.pe/}.
}
\details{
Available layer:
\itemize{
\item \bold{vegetation cover:} Polygons representing the agricultural areas of Peru, produced with high-resolution satellite images (RapidEye and Sentinel-2).
\item \bold{agriculture sector:} Polygons representing the new national register of agricultural statistical sectors for the year 2024.
\item \bold{oil palm:} Polygons representing the areas cultivated with oil palm in Peru for the period 2016 to 2020.
}
}
\examples{
\donttest{
library(geoidep)
library(sf)
# Disable the use of S2 geometry for accurate spatial operations
sf_use_s2(use_s2 = FALSE)

# Retrieve the polygon for Coronel Portillo province
coronel_portillo <- get_provinces(show_progress = FALSE) |>
  subset(NOMBPROV == "CORONEL PORTILLO") |>
  st_transform(crs = 32718)

# Download and extract the oil palm layer for Coronel Portillo
oil_palm <- get_midagri_data(
  layer = "oil palm",
  show_progress = FALSE
) |>
  st_intersection(coronel_portillo)

# Visualize the oil palm layer
plot(st_geometry(oil_palm))

}
}
