% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_midagri_link}
\alias{get_midagri_link}
\title{MIDAGRI links for obtaining cartographic information}
\usage{
get_midagri_link(type = NULL)
}
\arguments{
\item{type}{A string. Select only one of the following layers; 'vegetation cover', 'agriculture sector', 'oil palm'. Defaults to NULL.}
}
\value{
A string containing the URL of the requested file.
}
\description{
MIDAGRI links for obtaining cartographic information
}
\keyword{internal}
