% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mtc.R
\name{get_mtc_data}
\alias{get_mtc_data}
\title{Download the available data from MTC}
\usage{
get_mtc_data(
  layer = NULL,
  dsn = NULL,
  show_progress = TRUE,
  quiet = TRUE,
  timeout = 60
)
}
\arguments{
\item{layer}{Select only one from the list of available layers, for more information please use \code{get_data_sources(provider = "mtc")}. Defaults to NULL.}

\item{dsn}{Character. Output filename with the \strong{spatial format}. If missing, a temporary file is created.}

\item{show_progress}{Logical. Show a cli progress bar. Default \code{TRUE}.}

\item{quiet}{Logical. Suppress info message. Default \code{TRUE}.}

\item{timeout}{Numeric. Seconds to wait for a response. Default 60.}
}
\value{
An sf object.
}
\description{
Download the latest version of data available on the MTC geoportal.
For more information, visit \href{https://geoportal.mtc.gob.pe/}{MTC Geoportal}.
}
\examples{
\dontrun{
library(geoidep)
library(sf)
aerodromo <- get_mtc_data(layer = "aerodromos_2023", show_progress = FALSE)
head(aerodromo)
plot(st_geometry(aerodromo))
}
}
