% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_mtc_link}
\alias{get_mtc_link}
\title{Retrieve links to MTC for information on transport and telecomunication.}
\usage{
get_mtc_link(type = NULL)
}
\arguments{
\item{type}{A string. Select only one from the list of available layers, for more information please use `get_data_sources(provider = "mtc")`. Defaults to NULL.}
}
\value{
A string containing the URL of the requested file.
}
\description{
Retrieve links to MTC for information on transport and telecomunication.
}
\keyword{internal}
