% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_senamhi.R
\name{get_nro_emision}
\alias{get_nro_emision}
\title{Get information from a Senamhi weather alert according to the year.}
\usage{
get_nro_emision(nro = NULL)
}
\arguments{
\item{nro}{A numeric value representing the weather warning number.}
}
\value{
A tibble object.
}
\description{
This function obtains a table with the specified weather warning number.
For more information, please visit the following link: \url{https://www.Senamhi.gob.pe/?&p=aviso-meteorologico}
}
\examples{
\donttest{
library(geoidep)
get_nro_emision(295)
}
}
