% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/providers.R
\name{get_providers}
\alias{get_providers}
\title{List resources and count layers of IDEP}
\usage{
get_providers(query = NULL)
}
\arguments{
\item{query}{Character. Default is NULL (only \code{NULL} is valid).}
}
\value{
A tibble with columns \code{provider} and \code{layer_count}.
}
\description{
Summary of providers.
}
\examples{
\dontrun{
library(geoidep)
get_providers()
}
}
