% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_sernanp_link}
\alias{get_sernanp_link}
\title{Retrieve links to SERNANP for information on natural protected areas.}
\usage{
get_sernanp_link(type = NULL)
}
\arguments{
\item{type}{A string. Select only one from the list of available layers, for more information please use `get_data_sources(provider = "sernanp")`. Defaults to NULL.}
}
\value{
A string containing the URL of the requested file.
}
\description{
Retrieve links to SERNANP for information on natural protected areas.
}
\keyword{internal}
