% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_senamhi.R
\name{get_year}
\alias{get_year}
\title{Download Meteorological Alerts from Senamhi
This function downloads the weather forecast table provided by Senamhi according to the specified year.}
\usage{
get_year(year = NULL)
}
\arguments{
\item{year}{A numeric value indicating the year of publication of the weather warning.}
}
\value{
A tibble object.
}
\description{
Download Meteorological Alerts from Senamhi
This function downloads the weather forecast table provided by Senamhi according to the specified year.
}
\examples{
\donttest{
library(geoidep)
get_year(2024) |>
 head()
}
}
