% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{global-variables}
\alias{global-variables}
\title{Global variables for get_early_warning
This code declares global variables used in the `get_early_warning` function to avoid R CMD check warnings.}
\description{
Global variables for get_early_warning
This code declares global variables used in the `get_early_warning` function to avoid R CMD check warnings.
}
\keyword{internal}
