% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapbiomas-fire.R
\name{mapbiomas-fire}
\alias{mapbiomas-fire}
\title{MapBiomas Peru Fire (Fuego) rasters}
\description{
Annual / accumulated burned-area products from MapBiomas Fuego Peru
(Collection 1), hosted as GeoTIFFs on Google Cloud Storage and read
lazily via GDAL \verb{/vsicurl/}.
}
