% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/senamhi.R
\name{senamhi_alerts_by_year}
\alias{senamhi_alerts_by_year}
\title{Filter Senamhi alerts by year}
\usage{
senamhi_alerts_by_year(data, year)
}
\arguments{
\item{data}{A data frame of meteorological alerts.}

\item{year}{A numeric value indicating the year.}
}
\value{
A filtered tibble.
}
\description{
Filter Senamhi alerts by year
}
