% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/senamhi.R
\name{senamhi_get_meteorological_table}
\alias{senamhi_get_meteorological_table}
\title{Download Weather Alert Table from Senamhi}
\usage{
senamhi_get_meteorological_table(show_progress = TRUE, timeout = 60)
}
\arguments{
\item{show_progress}{Logical. Show a cli progress bar. Default \code{TRUE}.}

\item{timeout}{Numeric. Seconds to wait for the server response. Default 60.}
}
\value{
A tibble object containing the weather alert data.
}
\description{
This function downloads the table of weather warnings provided by Senamhi.
For more information, please visit the following link: \url{https://www.senamhi.gob.pe/?&p=aviso-meteorologico}
}
\examples{
\dontrun{
data <- senamhi_get_meteorological_table(show_progress = FALSE)
head(data)
}
}
