% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/senamhi.R
\name{senamhi_get_spatial_alerts}
\alias{senamhi_get_spatial_alerts}
\title{Download Meteorological Alert Geometry from Senamhi}
\usage{
senamhi_get_spatial_alerts(
  data = NULL,
  nro = NULL,
  year = NULL,
  dsn = NULL,
  show_progress = TRUE,
  quiet = TRUE,
  timeout = 60
)
}
\arguments{
\item{data}{A data frame with a single alert (with columns \code{nro} and \code{emision}), or NULL if you specify \code{nro} and \code{year}.}

\item{nro}{A numeric value (optional if \code{data} is provided).}

\item{year}{A numeric value (optional if \code{data} is provided).}

\item{dsn}{Path to save the downloaded .zip. If NULL, a temporary file is used.}

\item{show_progress}{Logical, show cli progress bar.}

\item{quiet}{Logical, suppress messages from \code{sf::st_read()}.}

\item{timeout}{Numeric. Seconds to wait for a response. Default 60.}
}
\value{
An sf object with the alert geometry.
}
\description{
This function downloads the spatial alert geometry (shapefile) from Senamhi for a given alert.
}
