% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serfor.R
\name{serfor}
\alias{serfor}
\title{SERFOR hotspots (SAMI)}
\description{
Active fire / heat-spot data from the Satellite Monitoring Unit (SAMI) of
the National Forestry and Wildlife Service:
\url{https://sniffs.serfor.gob.pe/monitoreo/sami/index.html}.
}
