% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoidep-package.R
\docType{package}
\name{geoidep-package}
\alias{geoidep}
\alias{geoidep-package}
\title{geoidep: Download Geographic Data on Various Topics Provided and Managed by the Spatial Data Infrastructure of Peru}
\description{
Provides R users with easy access to official cartographic data from Peru across a range of topics, including society, transport, environment, agriculture, climate, and more. It also includes data from regional government entities and technical-scientific institutions, all managed by Peru's Spatial Data Infrastructure. For more information, please visit: \url{https://www.geoidep.gob.pe/}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://geografo.pe/geoidep/data:image/png;base64},iVBORw0KGgoAAAANSUhEUgAAABIAAAASCAYAAABWzo5XAAAAbElEQVR4Xs2RQQrAMAgEfZgf7W9LAguybljJpR3wEse5JOL3ZObDb4x1loDhHbBOFU6i2Ddnw2KNiXcdAXygJlwE8OFVBHDgKrLgSInN4WMe9iXiqIVsTMjH7z/GhNTEibOxQswcYIWYOR/zAjBJfiXh3jZ6AAAAAElFTkSuQmCC
  \item Report bugs at \url{https://github.com/ambarja/geoidep/issues}
}

}
\author{
\strong{Maintainer}: antony barja \email{geografo.pe@gmail.com} (\href{https://orcid.org/0000-0001-5921-2858}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item antony barja \email{geografo.pe@gmail.com} (\href{https://orcid.org/0000-0001-5921-2858}{ORCID}) [copyright holder]
}

}
\keyword{internal}
