% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data_sources.R
\name{get_data_sources}
\alias{get_data_sources}
\title{List providers, description, year and link}
\usage{
get_data_sources()
}
\value{
A tibble object.
}
\description{
Metadata information
}
\examples{
\donttest{
library(geoidep)
dep <- get_departaments(show_progress = FALSE)
head(dep)
}
}
