% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data_sources.R
\name{get_data_sources}
\alias{get_data_sources}
\title{List providers, description, year and link}
\usage{
get_data_sources(query = NULL)
}
\arguments{
\item{query}{A string. Default is NULL. List of available providers. For more details, use the `get_providers` function.}
}
\value{
A tibble object.
}
\description{
This function allows you to list the list of available providers of the geoidep package.
}
\examples{
\donttest{
library(geoidep)
get_data_sources()
}
}
