% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inei.R
\name{get_departaments}
\alias{get_departaments}
\title{Download INEI departmental boundaries}
\usage{
get_departaments(
  departamento = NULL,
  dsn = NULL,
  show_progress = TRUE,
  quiet = TRUE,
  timeout = 60
)
}
\arguments{
\item{departamento}{Character. Name or names in a vector of the level-1 administrative boundary (department) to query. Case-insensitive.}

\item{dsn}{Character. Directory where the file will be saved. If \code{NULL}, a temporary file is used.}

\item{show_progress}{Logical. Show a \link[cli:cli_progress_bar]{cli} progress bar. Default \code{TRUE}.}

\item{quiet}{Logical. Suppress info message from \code{sf::st_read()}.}

\item{timeout}{Numeric. Seconds to wait for a response. Default 60.}
}
\value{
An sf object.
}
\description{
Download the latest version of the \strong{geometry} and \strong{ubigeos}
corresponding to the official political division of the departament boundaries of Peru.
For more information, visit \href{https://ide.inei.gob.pe/}{INEI Spatial Data Portal}.
}
\examples{
\dontrun{
library(geoidep)
dep <- get_departaments(show_progress = FALSE)
head(dep)
loreto <- get_departaments(departamento = "loreto", show_progress = FALSE)
head(loreto)
}
}
