% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geobosque.R
\name{get_early_warning}
\alias{get_early_warning}
\title{Download information on the latest deforestation alerts detected by Geobosque}
\usage{
get_early_warning(region, sf = TRUE, show_progress = TRUE)
}
\arguments{
\item{region}{An sf object. Area of interest (must be EPSG:4326).}

\item{sf}{Logical. Return an \code{sf} object (\code{TRUE}) or a tibble (\code{FALSE}).}

\item{show_progress}{Logical. Show cli progress. Default \code{TRUE}.}
}
\value{
A tibble or sf object.
}
\description{
Download deforestation alert information detected by Geobosque for any polygon in Peru.
For more details, visit \href{https://geobosques.minam.gob.pe}{Geobosque Platform}.
}
\examples{
\dontrun{
library(geoidep)
loreto <- get_departaments(show_progress = FALSE) |> subset(nombdep == "LORETO")
warning_point <- get_early_warning(region = loreto, sf = TRUE, show_progress = FALSE)
head(warning_point)
}
}
