% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_forest_loss_data.R
\name{get_forest_loss_data}
\alias{get_forest_loss_data}
\title{Download the forest and loss information from Geobosque}
\usage{
get_forest_loss_data(level = NULL, ubigeo = NULL, show_progress = TRUE)
}
\arguments{
\item{level}{A string. Specifies the official administrative division level of interest.}

\item{ubigeo}{A string Specifies the unique geographical code of interest.}

\item{show_progress}{Logical. Indicates whether to display the progress bar.}
}
\value{
A tibble object.
}
\description{
This function allows you to download the \bold{ubigeos} corresponding to the official political division
of the district, province or region boundaries of Peru with \bold{forest and loss information}.
For more information, you can visit the following website: \url{https://geobosques.minam.gob.pe}
}
\details{
Available layers are:
\itemize{
  \item \bold{dist:} A 6-character ubigeo representing the administrative boundaries of a district.
  \item \bold{prov:} A 4-character ubigeo representing the administrative boundaries of a province.
  \item \bold{dep:} A 2-character ubigeo representing the administrative boundaries of a region.
}
}
\examples{
\donttest{
library(geoidep)
geobosque <- get_forest_loss_data(ubigeo = "010101", level = "dist", show_progress = FALSE)
geobosque
}
}
