% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_geobosque_link}
\alias{get_geobosque_link}
\title{Geobosque API that returns data on forest stock, forest loss, forest loss by ranges for a given department, province and district.}
\usage{
get_geobosque_link(type = NULL)
}
\arguments{
\item{type}{A string. Select only one of the following layers; 'dist', 'pro', 'dep'. Defaults to NULL.}
}
\value{
A string containing the URL of the requested file.
}
\description{
Geobosque API that returns data on forest stock, forest loss, forest loss by ranges for a given department, province and district.
}
\keyword{internal}
