% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serfor.R
\name{get_hotspots_data}
\alias{get_hotspots_data}
\title{Download available hot spot data from Serfor's Satellite Monitoring Unit}
\usage{
get_hotspots_data(dsn = NULL, show_progress = TRUE, quiet = TRUE)
}
\arguments{
\item{dsn}{Character. Output filename. If missing, a temporary file is created.}

\item{show_progress}{Logical. Show a cli progress bar. Default \code{TRUE}.}

\item{quiet}{Logical. Suppress message from \code{sf::st_read()}. Default \code{TRUE}.}
}
\value{
An sf object.
}
\description{
Download the latest version of forest fire data available from the Satellite Monitoring Unit of the National Forestry and Wildlife Service of Peru.
For more information, visit \href{https://sniffs.serfor.gob.pe/monitoreo/sami/index.html}{Serfor Platform}.
}
\examples{
\dontrun{
library(geoidep)
library(sf)
hot_spots <- get_hotspots_data(show_progress = FALSE)
head(hot_spots)
}
}
