% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_inei_link}
\alias{get_inei_link}
\title{Retrieve INEI Links of the Basic Cartographic Information}
\usage{
get_inei_link(type = NULL)
}
\arguments{
\item{type}{String. One of 'districts', 'provinces', or 'departments'. Defaults to NULL.}
}
\value{
A string containing the URL of the requested file.
}
\description{
Retrieve INEI Links of the Basic Cartographic Information
}
\keyword{internal}
