% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mapbiomas_peru_fire_legend.R
\name{get_mapbiomas_peru_fire_legend}
\alias{get_mapbiomas_peru_fire_legend}
\title{Get the MapBiomas Peru Fire legend for a given product}
\usage{
get_mapbiomas_peru_fire_legend(product)
}
\arguments{
\item{product}{Character. One of the products listed in
\code{\link{get_mapbiomas_peru_fire_products}}.}
}
\value{
A tibble with columns \code{id}, \code{class_en}, \code{class_es},
and \code{hex}.
}
\description{
Returns a lookup table with the pixel value codes, English/Spanish class
names, and hexadecimal colors used by a given MapBiomas Fuego Peru
sub-product (Collection 1). Useful for building color scales (see
\code{\link{scale_fill_mapbiomas_peru_fire_d}}).
}
\details{
\itemize{
\item \code{annual_burned} and \code{accumulated_burned}: binary
classification (\code{1 = "Burned area"}).
\item \code{annual_burned_coverage} and \code{accumulated_burned_coverage}:
share the LULC legend (see \code{\link{get_mapbiomas_peru_legend}}),
since these products classify the burned pixels by land cover class.
\item \code{monthly_burned}: pixel values \code{1-12} represent the
month (January-December) in which the pixel burned.
\item \code{frequency_burned}: pixel values \code{1-12} represent the
number of times a pixel burned during the period (\code{12} = "12 or more times").
\item \code{year_last_fire}: pixel values are the actual years (\code{2013-2024})
of the most recent fire.
\item \code{annual_burned_scar_size_range}: pixel values \code{1-5}
represent burned scar size classes (from \code{< 25 ha} to \code{> 5000 ha}).
}
}
\examples{
library(geoidep)
get_mapbiomas_peru_fire_legend("frequency_burned")
}
