% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_midagri_data.R
\name{get_midagri_data}
\alias{get_midagri_data}
\title{Download the geographic information from MIDAGRI}
\usage{
get_midagri_data(dsn = NULL, layer = NULL, show_progress = TRUE, quiet = TRUE)
}
\arguments{
\item{dsn}{Character. The output filename in \bold{*.gpkg} format. If not provided, a temporary file will be created.}

\item{layer}{Character. Currently, only one layer is available.}

\item{show_progress}{Logical. If TRUE, displays a progress bar.}

\item{quiet}{Logical. If TRUE, suppresses informational messages.}
}
\value{
An `sf` object containing the downloaded geographic data.
}
\description{
This function downloads the latest version of the \bold{geographic data} representing the \bold{agricultural area} of Peru.
For more details, please visit the following page: \url{https://siea.midagri.gob.pe/portal/informativos/superficie-agricola-peruana}.
}
\details{
Available layer:
\itemize{
\item \bold{vegetation cover:} The agricultural area of Peru in shapefile format, provided at the national level by SIEA.
}
}
\examples{
\donttest{
library(geoidep)
library(sf)
# Disable the use of S2 geometry for spatial operations
sf_use_s2(use_s2 = FALSE)

# Downloading data and filtering by department
junin <- get_departaments(show_progress = FALSE) |>
  subset(NOMBDEP == "JUNIN")

# Performing an intersection between the Junin department and vegetation cover
cov_veg <- get_midagri_data(layer = "vegetation cover", show_progress = FALSE) |>
  st_intersection(junin)

# Plotting the geometry of the intersected vegetation cover
plot(st_geometry(cov_veg))
}
}
