% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_midagri_data.R
\name{get_midagri_data}
\alias{get_midagri_data}
\title{Download the geographic information from MIDAGRI}
\usage{
get_midagri_data(layer = NULL, dsn = NULL, show_progress = TRUE, quiet = TRUE)
}
\arguments{
\item{layer}{Character. Currently, Has three layer available, please review in details sections.}

\item{dsn}{Character. The output filename in \bold{*.shp} format. If not provided, a temporary file will be created.}

\item{show_progress}{Logical. If TRUE, displays a progress bar.}

\item{quiet}{Logical. If TRUE, suppresses informational messages.}
}
\value{
An `sf` object containing the downloaded geographic data.
}
\description{
This function downloads the latest version of the \bold{geographic data} representing the \bold{agricultural area} of Peru.
For more details, please visit the following page: \url{https://siea.midagri.gob.pe/portal/informativos/superficie-agricola-peruana}.
}
\details{
Available layer:
\itemize{
\item \bold{vegetation cover:} Polygons representing the agricultural areas of Peru, produced with high-resolution satellite images (RapidEye and Sentinel-2).
\item \bold{agriculture sector:} Polygons representing the new national register of agricultural statistical sectors for the year 2024.
\item \bold{oil palm:} Polygons representing the areas cultivated with oil palm in Peru for the period 2016 to 2020.
}
}
\examples{
\donttest{
library(geoidep)
library(mapview)
oil_palm <- get_midagri_data(layer = "oil palm", show_progress = FALSE)
# Plotting the geometry of the vegetation cover
mapview(oil_palm)
}
}
