% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_providers.R
\name{get_providers}
\alias{get_providers}
\title{List resources and count layers of IDEP}
\usage{
get_providers(query = "provider")
}
\arguments{
\item{query}{description}
}
\value{
A sf or tibble object.
}
\description{
Summary of providers
}
\examples{
\donttest{
library(geoidep)
get_providers()
}
}
