% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inei.R
\name{get_provinces}
\alias{get_provinces}
\title{Download INEI province boundaries}
\usage{
get_provinces(
  departamento = NULL,
  provincia = NULL,
  dsn = NULL,
  show_progress = TRUE,
  quiet = TRUE,
  timeout = 60
)
}
\arguments{
\item{departamento}{Character. Name of the level-1 administrative boundary (department). Case-insensitive.}

\item{provincia}{Character. Name of the level-2 administrative boundary (province). Case-insensitive. Requires \code{departamento}.}

\item{dsn}{Character. Directory where the file will be saved. If \code{NULL}, a temporary file is used.}

\item{show_progress}{Logical. Show a \link[cli:cli_progress_bar]{cli} progress bar.}

\item{quiet}{Logical. Suppress info message.}

\item{timeout}{Numeric. Seconds to wait for a response. Default 60.}
}
\value{
An sf object.
}
\description{
Download the latest version of the \strong{geometry} and \strong{ubigeos}
corresponding to the official political division of the province boundaries of Peru.
For more information, visit \href{https://ide.inei.gob.pe/}{INEI Spatial Data Portal}.
}
\examples{
\dontrun{
library(geoidep)
prov <- get_provinces(show_progress = FALSE)
head(prov)
}
}
