% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_provinces.R
\name{get_provinces}
\alias{get_provinces}
\title{Download INEI province boundaries}
\usage{
get_provinces(
  departamento = NULL,
  provincia = NULL,
  dsn = NULL,
  show_progress = TRUE,
  quiet = TRUE,
  timeout = 60
)
}
\arguments{
\item{departamento}{Character. Name of the level-1 administrative boundary (department) to query. The input is case-insensitive.}

\item{provincia}{Character. Name of the level-2 administrative boundary (province) to query. The input is case-insensitive.}

\item{dsn}{Character. Path to the output \code{.gpkg} file or a directory where the file will be saved.
If a directory is provided, the file will be saved as \code{provincia.gpkg} inside it.
If \code{NULL}, a temporary file will be created.
If the path contains multiple subdirectories, they will be created automatically if they do not exist.}

\item{show_progress}{Logical. Suppress bar progress.}

\item{quiet}{Logical. Suppress info message.}

\item{timeout}{Seconds. Number of seconds to wait for a response until giving up. Cannot be less than 1 ms. Default is 60.}
}
\value{
An sf or tibble object.
}
\description{
This function allows you to download the latest version of the \bold{geometry} and \bold{ubigeos}
corresponding to the \bold{official political division} of the province boundaries of Peru.
For more information, you can visit the following page \href{https://ide.inei.gob.pe/}{INEI Spatial Data Portal}.
}
\examples{
\donttest{
library(geoidep)
prov <- get_provinces(show_progress = FALSE)
head(prov)
}
}
