% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sernanp.R
\name{get_sernanp_data}
\alias{get_sernanp_data}
\title{Download the available data from Sernanp}
\usage{
get_sernanp_data(
  layer = NULL,
  dsn = NULL,
  show_progress = TRUE,
  quiet = TRUE,
  timeout = 60
)
}
\arguments{
\item{layer}{Select only one from the list of available layers, for more information please use \code{get_data_sources(provider = "sernanp")}. Defaults to NULL.}

\item{dsn}{Character. Output filename. If missing, a temporary file is created.}

\item{show_progress}{Logical. Show a cli progress bar. Default \code{TRUE}.}

\item{quiet}{Logical. Suppress info message. Default \code{TRUE}.}

\item{timeout}{Numeric. Seconds to wait for a response. Default 60.}
}
\value{
An sf object.
}
\description{
Download the latest version of data available on the sernanp geoviewer.
For more information, visit \href{https://geo.sernanp.gob.pe/visorsernanp/}{Sernanp Platform}.
}
\examples{
\dontrun{
library(geoidep)
library(sf)
anp <- get_sernanp_data(layer = "zonificacion_anp", show_progress = FALSE)
plot(st_geometry(anp))
}
}
