% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/senamhi.R
\name{senamhi_alert_by_number}
\alias{senamhi_alert_by_number}
\title{Filter Senamhi alerts by number}
\usage{
senamhi_alert_by_number(data, nro)
}
\arguments{
\item{data}{A data frame of meteorological alerts.}

\item{nro}{A numeric vector indicating the alert number(s).}
}
\value{
A filtered tibble.
}
\description{
Filter Senamhi alerts by number
}
