% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_senamhi.R
\name{senamhi_alerts_by_year}
\alias{senamhi_alerts_by_year}
\title{Download Senamhi's Meteorological Alerts by year}
\usage{
senamhi_alerts_by_year(data = NULL, year = NULL)
}
\arguments{
\item{data}{A tibble object containing the weather alert data.}

\item{year}{A numeric value indicating the year of publication of the weather warning.}
}
\description{
This function downloads the weather forecast table provided by Senamhi according to the specified year.
}
\examples{
\donttest{
library(geoidep)
data <- senamhi_alerts_by_year(year = 2024)
head(data)
}
}
