% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_senamhi.R
\name{senamhi_alerts_by_year}
\alias{senamhi_alerts_by_year}
\title{Download Senamhi's Meteorological Alerts by year
This function downloads the weather forecast table provided by Senamhi according to the specified year.}
\usage{
senamhi_alerts_by_year(year = NULL)
}
\arguments{
\item{year}{A numeric value indicating the year of publication of the weather warning.}
}
\description{
Download Senamhi's Meteorological Alerts by year
This function downloads the weather forecast table provided by Senamhi according to the specified year.
}
