% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_senamhi.R
\name{senamhi_geometry_by_level}
\alias{senamhi_geometry_by_level}
\title{Filter alert geometry by danger level}
\usage{
senamhi_geometry_by_level(sf_data, level)
}
\arguments{
\item{sf_data}{An sf object returned by \code{senamhi_get_spatial_alerts()}.}

\item{level}{Numeric or character vector: values like 1, 2, 3, 4 or "Nivel 1", etc.}
}
\value{
Filtered sf object.
}
\description{
This function filters the \code{sf} object from Senamhi by alert level (e.g., 1 to 4).
}
\examples{
\donttest{
senamhi_get_meteorological_table() |>
  senamhi_alert_by_number(295) |>
  senamhi_alerts_by_year(2024) |>
  senamhi_get_spatial_alerts() |>
  senamhi_geometry_by_level(3) |>
  plot()
}
}
