% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_senamhi.R
\name{senamhi_get_meteorological_table}
\alias{senamhi_get_meteorological_table}
\title{Download Weather Alert Table from Senamhi}
\usage{
senamhi_get_meteorological_table()
}
\value{
A tibble object containing the weather alert data.
}
\description{
This function downloads the table of weather warnings provided by Senamhi.
For more information, please visit the following link: \url{https://www.Senamhi.gob.pe/?&p=aviso-meteorologico}
}
\examples{
\donttest{
data <- senamhi_get_meteorological_table()
head(data)
}
}
